+++ src/index.css
@@ -3,7 +3,7 @@
 @tailwind utilities;
 
 .opt-container {
-  @apply flex justify-between gap-2 bg-blue-200;
+  @apply flex justify-between gap-2;
 }
 .opt-container input {
   @apply !mr-0 flex-1 py-5 outline-none;
