+++ src/pages/Error.jsx
@@ -1,0 +1,9 @@
+const Error = () => {
+  return (
+    <div className="p-4">
+      <p>you do not have permission to access or page not found!</p>
+    </div>
+  )
+}
+
+export default Error
